requirements_prod.txt
# ==============================================================================
# ATLAS Pro - Dependencias de Produccion
# ==============================================================================

# Core RL
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# API y Dashboard
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0

# NTCIP (SNMP)
pysnmp==4.4.12
pysnmp-mibs>=0.1.6

# HTTP client (UTMC, OTA)
requests>=2.31.0
lxml>=4.9.0

# ONNX (edge deployment)
onnxruntime>=1.16.0

# Base de datos (metricas)
psycopg2-binary>=2.9.0

# Monitoreo
prometheus-client>=0.17.0

# Utilidades
python-dateutil>=2.8.0

# Dashboard v2 (Auth + Reportes)
python-multipart>=0.0.6
reportlab>=4.0.0
